refactor MinterSetPrice (ETH) tests
